Module kempt::map

Types supporting the Map<Key, Value> collection type.

Structs

Difference

An iterator over the difference between two Maps.

Drain

An iterator that removes all of the Fields of an Map.

EntryRef

A reference to a key from a Map and an associated value.

Field

A field in an Map.

Intersection

An iterator that yields entries that appear in two maps.

IntoIter

An iterator that returns all of the elements of an Map while freeing its underlying memory.

IntoKeys

An iterator converting a Map into a series of owned keys.

IntoValues

An iterator returning all of the values contained in an Map as its underlying storage is freed.

Iter

An iterator over the Fields in an Map.

IterMut

An iterator over mutable Fields contained in an Map.

Keys

An iterator over the keys in an Map.

Map

An ordered Key/Value map.

OccupiedEntry

An entry that exists in an Map.

Union

An iterator that yields Unioned entries for two Maps.

VacantEntry

A vacant entry in an Map.

Values

An iterator over the values contained in an Map.

ValuesMut

An iterator over mutable values contained in an Map.

Enums

Entry

The result of looking up an entry by its key.

OwnedOrRef

An owned value or a reference to a value of that type.

SearchKey

A key provided to the Map::entry function.

Unioned

A unioned entry from a Union iterator. An entry can be from the left, right, or both maps.