Struct kempt::set::Set

pub struct Set<T>(_)
where
         T: Ord;

An ordered collection of unique Ts.

This data type only allows each unique value to be stored once.

use kempt::Set;

let mut set = Set::new();
set.insert(1);
assert!(!set.insert(1));
assert_eq!(set.len(), 1);

The values in the collection are automatically sorted using T’s Ord implementation.

use kempt::Set;

let mut set = Set::new();
set.insert(1);
set.insert(3);
set.insert(2);
assert_eq!(set.member(0), Some(&1));
assert_eq!(set.member(1), Some(&2));
assert_eq!(set.member(2), Some(&3));

Implementations

impl<T> Set<T>where T: Ord,

pub const fn new() -> Self

Returns an empty set.

pub fn with_capacity(capacity: usize) -> Self

Returns an empty set with enough allocated memory to store capacity values without reallocating.

pub fn insert(&mut self, value: T) -> bool

Inserts or replaces value in the set, returning true if the collection is modified. If a previously contained value returns Ordering::Equal from Ord::cmp, the collection will not be modified and false will be returned.

pub fn replace(&mut self, value: T) -> Option<T>

Inserts or replaces value in the set. If a previously contained value returns Ordering::Equal from Ord::cmp, the new value will overwrite the stored value and it will be returned.

pub fn contains<SearchFor>(&self, value: &SearchFor) -> boolwhere T: Sort<SearchFor>, SearchFor: ?Sized,

Returns true if the set contains a matching value.

pub fn get<SearchFor>(&self, value: &SearchFor) -> Option<&T>where T: Sort<SearchFor>, SearchFor: ?Sized,

Returns the contained value that matches value.

pub fn remove<SearchFor>(&mut self, value: &SearchFor) -> Option<T>where T: Sort<SearchFor>, SearchFor: ?Sized,

Removes a value from the set, returning the value if it was removed.

pub fn member(&self, index: usize) -> Option<&T>

Returns the member at index inside of this ordered set. Returns None if index is greater than or equal to the set’s length.

pub fn len(&self) -> usize

Returns the number of members in this set.

pub fn is_empty(&self) -> bool

Returns true if there are no members in this set.

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator over the members in this set.

pub fn union<'a>(&'a self, other: &'a Set<T>) -> Union<'a, T>

Returns an iterator that yields a single reference to all members found in either self or other.

This iterator is guaranteed to return results in the sort order of the Key type.

pub fn intersection<'a>(&'a self, other: &'a Set<T>) -> Intersection<'a, T>

Returns an iterator that yields a single reference to all members found in both self and other.

This iterator is guaranteed to return results in the sort order of the Key type.

pub fn difference<'a>(&'a self, other: &'a Set<T>) -> Difference<'a, T>

Returns an iterator that yields a single reference to all members found in self but not other.

This iterator is guaranteed to return results in the sort order of the Key type.